requirements.txt
pandas==2.2.3
jupyterlab
jupytext
openai
keyboard
python-dotenv
openpyxl==3.1.5
fastapi==0.115.11
uvicorn==0.34.0
pydantic==2.9.2
pydantic-settings==2.6.1
pymongo==4.13.0
motor==3.7.1
requests==2.32.3
cryptography==43.0.3
PyJWT==2.10.1
python-multipart
